Separated formDraft create & update handlers
